code/aiResume/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
black==23.10.1
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
dependency-injector==4.41.0
Django==4.2.6
djangorestframework==3.14.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.0
openai==0.28.1
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
ulid-py==1.1.0
urllib3==2.0.7
yarl==1.9.2
